requirements_image.txt
transformers
datasets
accelerate

sentencepiece
streamlit

speedtest-cli